refactor: move encodeSchema from cli to schema-type package
